refactor(ui-rewrite): replace mcp server details Sheet with in-flow panel
Swaps the Radix Dialog/Sheet-based drawer for an in-flow panel scoped to
the SidebarInset main region. The sidebar and header stay interactive
and unblurred while the panel is open, matching the Figma design.

The panel uses role=region (not dialog), manages focus via refs and
effects, and handles ESC and backdrop click directly instead of relying
on Dialog's modal contract. No changes to sheet.tsx, sidebar, or other
drawers.
